+++ packages/plugins/didkit/CHANGELOG.md
@@ -1,5 +1,21 @@
 # learn-card-core
 
+## 1.4.1
+
+### Patch Changes
+
+-   [#388](https://github.com/learningeconomy/LearnCard/pull/388) [`336876b`](https://github.com/learningeconomy/LearnCard/commit/336876b4b98e37157b8a133ed3b72801eb3d1cd8) Thanks [@TaylorBeeston](https://github.com/TaylorBeeston)! - Emit declarationMap
+
+-   Updated dependencies [[`336876b`](https://github.com/learningeconomy/LearnCard/commit/336876b4b98e37157b8a133ed3b72801eb3d1cd8)]:
+    -   @learncard/types@5.3.4
+    -   @learncard/core@9.1.2
+
+## 1.4.0
+
+### Minor Changes
+
+-   [#387](https://github.com/learningeconomy/LearnCard/pull/387) [`f4d07f1`](https://github.com/learningeconomy/LearnCard/commit/f4d07f1559308b3a91de67d43197073e35699163) Thanks [@TaylorBeeston](https://github.com/TaylorBeeston)! - Do not error if Status List VC has a different issuer than the VC being verified
+
 ## 1.3.2
 
 ### Patch Changes