+++ packages/plugins/learn-cloud/CHANGELOG.md
@@ -1,5 +1,25 @@
 # learn-card-core
 
+## 1.1.25
+
+### Patch Changes
+
+-   [#388](https://github.com/learningeconomy/LearnCard/pull/388) [`336876b`](https://github.com/learningeconomy/LearnCard/commit/336876b4b98e37157b8a133ed3b72801eb3d1cd8) Thanks [@TaylorBeeston](https://github.com/TaylorBeeston)! - Emit declarationMap
+
+-   Updated dependencies [[`336876b`](https://github.com/learningeconomy/LearnCard/commit/336876b4b98e37157b8a133ed3b72801eb3d1cd8)]:
+    -   @learncard/learn-cloud-client@1.1.25
+    -   @learncard/helpers@1.0.9
+    -   @learncard/core@9.1.2
+    -   @learncard/didkit-plugin@1.4.1
+
+## 1.1.24
+
+### Patch Changes
+
+-   Updated dependencies [[`f4d07f1`](https://github.com/learningeconomy/LearnCard/commit/f4d07f1559308b3a91de67d43197073e35699163)]:
+    -   @learncard/didkit-plugin@1.4.0
+    -   @learncard/learn-cloud-client@1.1.24
+
 ## 1.1.23
 
 ### Patch Changes